feat: add login queries(#109)
